rename auth_utils to auth, and lazy load boto3 in it
